requirements.txt
azure-identity
azure-monitor-query
pandas
msal
openai
colorama
python-dotenv
requests
beautifulsoup4
msgraph-sdk
#Webapp
flask
flask-socketio